pr/develop

+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/mypage/MypageFragment.kt
@@ -5,14 +5,17 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
+import com.nbcfinalteam2.ddaraogae.R
 import com.nbcfinalteam2.ddaraogae.databinding.FragmentMypageBinding
 import com.nbcfinalteam2.ddaraogae.presentation.model.DefaultEvent
+import com.nbcfinalteam2.ddaraogae.presentation.ui.dog.MyPetActivity
 import com.nbcfinalteam2.ddaraogae.presentation.ui.add.AddActivity
-import com.nbcfinalteam2.ddaraogae.presentation.ui.dog.DetailPetActivity
+import com.nbcfinalteam2.ddaraogae.presentation.ui.alarm.AlarmActivity
 import com.nbcfinalteam2.ddaraogae.presentation.ui.loading.LoadingDialog
 import com.nbcfinalteam2.ddaraogae.presentation.util.ToastMaker
 import dagger.hilt.android.AndroidEntryPoint
@@ -41,25 +44,38 @@ class MypageFragment : Fragment() {
         clickAboutAccountBtn()
         clickAboutPetBtn()
         clickPrivacyBtn()
+        clickAboutFunctionBtn()
         initViewModel()
 
     }
 
     private fun clickAboutAccountBtn(){
         binding.tvSignOut.setOnClickListener {
-            viewModel.logOut()
+            val builder = AlertDialog.Builder(requireContext())
+            builder.setMessage(R.string.msg_logout)
+            builder.setPositiveButton(R.string.mypage_delete_dog_thumbnail_positive) { _, _ ->
+                viewModel.logOut()
+            }
+            builder.setNegativeButton(R.string.mypage_delete_dog_thumbnail_negative) { _, _ -> }
+            builder.show()
         }
 
         binding.tvSignDelete.setOnClickListener {
-            viewModel.deleteUser()
+            val builder = AlertDialog.Builder(requireContext())
+            builder.setTitle(R.string.msg_delete_account)
+            builder.setMessage(R.string.msg_delete_account_context)
+            builder.setPositiveButton(R.string.mypage_delete_dog_thumbnail_positive) { _, _ ->
+                viewModel.deleteUser()
+            }
+            builder.setNegativeButton(R.string.mypage_delete_dog_thumbnail_negative) { _, _ -> }
+            builder.show()
+
         }
     }
     private fun clickAboutPetBtn(){
-        binding.tvMyDogAdd.setOnClickListener {
-            startActivity(Intent(requireActivity(), AddActivity::class.java))
-        }
+
         binding.tvMyDogEdit.setOnClickListener {
-            startActivity(Intent(requireActivity(), DetailPetActivity::class.java))
+            startActivity(Intent(requireActivity(), MyPetActivity::class.java))
         }
     }
     private fun clickPrivacyBtn(){
@@ -71,6 +87,11 @@ class MypageFragment : Fragment() {
         }
         binding.tvAgreementPrivacyPolicy.setOnClickListener {
             startActivity(Intent(requireActivity(), MypageAgreementPrivacy::class.java))
+        }
+    }
+    private fun clickAboutFunctionBtn() {
+        binding.tvSetAlarm.setOnClickListener {
+            startActivity(Intent(requireActivity(), AlarmActivity::class.java))
         }
     }
